+++ src/njoy21/input/PLOTR/Card10a/Xtag/test/Xtag.test.cpp
@@ -11,7 +11,7 @@ using namespace njoy::njoy21::input;
 SCENARIO( "Xtag output values", "[PLOTR],[Card10a], [Xtag]"){
 
   GIVEN( "valid Xtag parameters" ){
-    std::vector<float> validValues{0.0,0.1,0.5,0.8,1.0,2.0,80.0};
+    std::vector<double> validValues{0.0,0.1,0.5,0.8,1.0,2.0,80.0};
 
     THEN( "the returned class has the correct value" ){
       for( auto xtag : validValues ){
@@ -25,7 +25,7 @@ SCENARIO( "Xtag output values", "[PLOTR],[Card10a], [Xtag]"){
   } // GIVEN
 
   GIVEN( "invalid Xtag parameters" ){
-    std::vector<float> invalidValues{ -2.0, -1.1 };
+    std::vector<double> invalidValues{ -2.0, -1.1 };
 
     THEN( "the class throws an exception" ){
       for( auto xtag : invalidValues ){
