+++ src/njoy21/input/PLOTR/Card11/Zv.hpp
@@ -1,5 +1,5 @@
 struct Zv {
-  using Value_t = float;
+  using Value_t = double;
   static std::string name(){ return "zv"; }
   static std::string description(){
     return
