+++ src/njoy21/input/PLOTR/Card2/Wh.hpp
@@ -1,5 +1,5 @@
 struct Wh {
-  using Value_t = float;
+  using Value_t = double;
   static std::string name(){ return "wh"; }
   static std::string description(){
     return
